reference/qa/requirements.txt
torch==1.6.0+cpu
transformers==3.1.0
grpcio
protobuf